+++ ci/release_notes.md
@@ -2,3 +2,6 @@
 
 - pick up `RISQ_HOME` as root data dir
 - Expose / remove code when `target_os = "android"`
+
+## Bug fixes
+- Add capability to handle new TradeStatistics hash